Fix printing of filenames with surrogate escapes
On Unix systems a filename can be a sequence of bytes that is not valid
UTF-8. Python uses[1] surrogate escapes to allow to decode such
filenames to Unicode (bytes that cannot be decoded are replaced by a
surrogate; upon encoding the surrogate is converted to the original
byte).

From `click` docs[2]:

> Invalid bytes or surrogate escapes will raise an error when written
> to a stream with `errors="strict"`. This will typically happen with
> `stdout` when the locale is something like `en_GB.UTF-8`.

To fix that, we use `click.format_filename`[2] before printing the
filenames to `stdout` so that surrogate escapes are replaced by �.

Fixes #768

[1]: https://peps.python.org/pep-0383/
[2]: https://click.palletsprojects.com/en/8.1.x/api/#click.format_filename